Refactor to own class
